requirements.txt
Django==1.11.3
djangorestframework==3.6.3
olefile==0.45.1
Pillow==4.3.0
pytz==2018.4
